Resized when bounds change
Didn't realise this when working on android but the bottom sheet size needs to match the native size. So it should grow when it's being dragged. Although the code is copied from react native modal it's not working - it resizes properly when dragged but if drag a bit and let go then there's a gap at the bottom as the sheet animates back. No idea why viewDidLayoutSubviews isn't called with the intermediate frame sizes. Checking this in so can show to jacob parker - see if he knows why